HKUST/COMP4651/project/face-blur/function/requirements.txt
opencv-python-headless
mtcnn-opencv
numpy
minio
Pillow